pyproject.toml
[project]
name = "rag-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "fastapi==0.104.1",
    "numpy>=2.3.2",
    "openai>=1.100.2",
    "pydantic>=2.11.7",
    "pymongo>=4.14.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "tiktoken>=0.11.0",
    "uvicorn>=0.35.0",
]
